fix(frontend): clear dead export and Sonar duplication
Remove unused normalizeToStartOfDay and share billing-period filter
plus chart today/transform helpers to drop new-code duplication.
