+++ src/components/dashboard/DeploymentStatsCard.vue
@@ -14,6 +14,7 @@ import {
 import { useSupabase } from '~/services/supabase'
 import { useDashboardAppsStore } from '~/stores/dashboardApps'
 import { useOrganizationStore } from '~/stores/organization'
+import { filterDailySeriesToBillingPeriod } from '~/utils/chartOptimizations'
 import ChartCard from './ChartCard.vue'
 import DeploymentStatsChart from './DeploymentStatsChart.vue'
 
@@ -39,43 +40,6 @@ const props = defineProps({
     default: false,
   },
 })
-
-// Helper function to filter 30-day data to billing period
-function filterToBillingPeriod(fullData: number[], last30DaysStart: Date, billingStart: Date) {
-  const currentDate = normalizeToUtcStartOfDay()
-
-  // Calculate billing period length
-  let currentBillingDay: number
-
-  if (billingStart.getUTCDate() === 1) {
-    currentBillingDay = currentDate.getUTCDate()
-  }
-  else {
-    const billingStartDay = billingStart.getUTCDate()
-    const daysInMonth = new Date(Date.UTC(currentDate.getUTCFullYear(), currentDate.getUTCMonth() + 1, 0)).getUTCDate()
-    currentBillingDay = (currentDate.getUTCDate() - billingStartDay + 1 + daysInMonth) % daysInMonth
-    if (currentBillingDay === 0)
-      currentBillingDay = daysInMonth
-  }
-
-  // Create arrays for billing period length
-  const billingData = Array.from({ length: currentBillingDay }).fill(0) as number[]
-
-  // Map 30-day data to billing period
-  for (let i = 0; i < 30; i++) {
-    const dataDate = addUtcDays(last30DaysStart, i)
-
-    // Check if this date falls within current billing period
-    if (dataDate >= billingStart && dataDate <= currentDate) {
-      const billingIndex = Math.floor((dataDate.getTime() - billingStart.getTime()) / (1000 * 60 * 60 * 24))
-      if (billingIndex >= 0 && billingIndex < currentBillingDay) {
-        billingData[billingIndex] = fullData[i]
-      }
-    }
-  }
-
-  return { data: billingData }
-}
 
 const { t } = useI18n()
 const organizationStore = useOrganizationStore()
@@ -305,19 +269,19 @@ async function calculateStats(forceRefetch = false) {
     let finalTotal = totalDeploymentsCount
 
     if (props.useBillingPeriod) {
-      const filteredData = filterToBillingPeriod(dailyCounts30Days, last30DaysStart, billingStart)
+      const filteredData = filterDailySeriesToBillingPeriod(dailyCounts30Days, last30DaysStart, billingStart)
       finalDeploymentData = filteredData.data
 
       const filteredPerChannel: { [channelId: string]: number[] } = {}
       Object.keys(perChannel).forEach((channelId) => {
-        const filteredChannelData = filterToBillingPeriod(perChannel[channelId], last30DaysStart, billingStart)
+        const filteredChannelData = filterDailySeriesToBillingPeriod(perChannel[channelId], last30DaysStart, billingStart)
         filteredPerChannel[channelId] = filteredChannelData.data
       })
       finalPerChannel = filteredPerChannel
 
       const filteredPerApp: { [appId: string]: number[] } = {}
       Object.keys(perApp).forEach((appId) => {
-        const filteredAppData = filterToBillingPeriod(perApp[appId], last30DaysStart, billingStart)
+        const filteredAppData = filterDailySeriesToBillingPeriod(perApp[appId], last30DaysStart, billingStart)
         filteredPerApp[appId] = filteredAppData.data
       })
       finalPerApp = filteredPerApp
