+++ src/components/dashboard/DeploymentStatsChart.vue
@@ -17,6 +17,7 @@ import { computed } from 'vue'
 import { Bar, Line } from 'vue-chartjs'
 import { useI18n } from 'vue-i18n'
 import { useRouter } from 'vue-router'
+import { getTodayLimit as resolveTodayLimit, transformSeries as transformDailySeries } from '~/services/buildCharts'
 import { createLegendConfig, createStackedChartScales } from '~/services/chartConfig'
 import { createTooltipConfig, todayLinePlugin, verticalLinePlugin } from '~/services/chartTooltip'
 import { generateMonthDays, getDaysInCurrentMonth, normalizeToUtcStartOfDay } from '~/services/date'
@@ -45,7 +46,13 @@ const organizationStore = useOrganizationStore()
 const cycleStart = normalizeToUtcStartOfDay(new Date(organizationStore.currentOrganization?.subscription_start ?? new Date()))
 const cycleEnd = normalizeToUtcStartOfDay(new Date(organizationStore.currentOrganization?.subscription_end ?? new Date()))
 
-const DAY_IN_MS = 1000 * 60 * 60 * 24
+function getTodayLimit(labelCount: number) {
+  return resolveTodayLimit(labelCount, props.useBillingPeriod, cycleStart, cycleEnd)
+}
+
+function transformSeries(source: number[], accumulated: boolean, labelCount: number) {
+  return transformDailySeries(source, accumulated, labelCount, getTodayLimit(labelCount))
+}
 
 // Determine mode based on which data is provided
 const isChannelMode = computed(() => Object.keys(props.dataByChannel).length > 0)
@@ -102,51 +109,6 @@ Chart.register(
   CategoryScale,
   LinearScale,
 )
-
-function getTodayLimit(labelCount: number) {
-  if (!props.useBillingPeriod)
-    return labelCount - 1
-
-  const today = normalizeToUtcStartOfDay()
-
-  // If cycle end is today or in the past, show all data
-  if (cycleEnd <= today)
-    return labelCount - 1
-
-  // If cycle end is in the future, only show data up to today
-  const diff = Math.floor((today.getTime() - cycleStart.getTime()) / DAY_IN_MS)
-
-  if (Number.isNaN(diff) || diff < 0)
-    return -1
-
-  return Math.min(diff, labelCount - 1)
-}
-
-function transformSeries(source: number[], accumulated: boolean, labelCount: number) {
-  const display: Array<number | null> = Array.from({ length: labelCount }).fill(null) as Array<number | null>
-  const base: Array<number | null> = Array.from({ length: labelCount }).fill(null) as Array<number | null>
-  const limitIndex = getTodayLimit(labelCount)
-
-  if (limitIndex < 0)
-    return { display, base }
-
-  let runningTotal = 0
-  for (let index = 0; index <= limitIndex; index++) {
-    const hasValue = index < source.length && typeof source[index] === 'number' && Number.isFinite(source[index])
-    const numericValue = hasValue ? source[index] as number : 0
-
-    base[index] = numericValue
-    if (accumulated) {
-      runningTotal += numericValue
-      display[index] = runningTotal
-    }
-    else {
-      display[index] = numericValue
-    }
-  }
-
-  return { display, base }
-}
 
 function monthdays() {
   return generateMonthDays(props.useBillingPeriod, cycleStart, cycleEnd)
